chaos.
we cannot compile traversal code with rta::simple_trinagle, as it
constructs vec3f for which we cannot (and don't want to) create
__host__ __device__ versions.
therefore we requrie different triangle types, which is what all the
templates were introduced to do. But the problems are a little deeper
as we also want to adapt existing code.
current state is a little chaotic as things still need some sorting
out...

see bbvh-cuda/main.cpp